Modify behaviour of 'Archiver.move_file()'
Modify 'move_file()' so that it accepts both an input file path and an output file path. This makes its behaviour more explicit and clear as well as being more flexible (previously it was impossible to rename the file during the move).

It can easily be modified to work more like the Linux 'mv' command, where it will use the existing file name if the target path is a directory but there is no need of this right now.